feat: add some extra process when dealing with DOM Event has the structure is hiding its properties
Dom Event (and all it sub classes) is built in a way it's internal properties are accessible when querying them directly but "hidden" when iterating its keys.

With a code example it means: `Object.keys(new Event('click')) = ["isTrusted"]`

So to be able to stringify/parse more than just `isTrusted` info we need to create a new object and set the properties by hand. As there is no way to iterate the properties we rely on a list of hardcoded properties.